Moving display adapter fn out of generated classes
